+++ ncrf/_typing.py
@@ -15,4 +15,5 @@ TrialData = tuple[FloatArray, FloatArray]
 ObjectiveFunction = Callable[[FloatArray], float]
 GradientFunction = Callable[[FloatArray], FloatArray]
 MuArg = float | Sequence[float] | FloatArray | Literal["auto"]
+ScaleArg = Literal["l1", "l2", "spectral"] | None
 StimDimensions = Categorial | Scalar | Space